% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{assumptions.dfr_series_md}
\alias{assumptions.dfr_series_md}
\title{Assumptions for masked-cause DFR series systems}
\usage{
\method{assumptions}{dfr_series_md}(model, ...)
}
\arguments{
\item{model}{A \code{\link{dfr_series_md}} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Character vector of model assumptions.
}
\description{
Assumptions for masked-cause DFR series systems
}
