% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_model_defaults}
\alias{extract_model_defaults}
\title{Extract model column name defaults}
\usage{
extract_model_defaults(model)
}
\arguments{
\item{model}{likelihood model object}
}
\value{
list with lifetime, lifetime_upper, omega, candset defaults
}
\description{
Extract model column name defaults
}
\keyword{internal}
