% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_series_md.R
\name{is_dfr_series_md}
\alias{is_dfr_series_md}
\title{Test whether an object is a dfr_series_md}
\usage{
is_dfr_series_md(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a dfr_series_md
}
\examples{
model <- dfr_series_md(components = list(
  dfr_exponential(0.1), dfr_exponential(0.2)
))
is_dfr_series_md(model)    # TRUE
is_dfr_series_md(42)       # FALSE
}
