% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ncomponents.dfr_series_md}
\alias{ncomponents.dfr_series_md}
\title{Number of components in a masked-cause DFR series system}
\usage{
\method{ncomponents}{dfr_series_md}(x, ...)
}
\arguments{
\item{x}{A \code{\link{dfr_series_md}} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Integer, the number of components.
}
\description{
Number of components in a masked-cause DFR series system
}
\examples{
model <- dfr_series_md(components = list(
  dfr_exponential(0.1), dfr_exponential(0.2), dfr_exponential(0.3)
))
ncomponents(model)   # 3
}
