% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{loglik}
\alias{score}
\alias{hess_loglik}
\alias{assumptions}
\alias{rdata}
\alias{fit}
\alias{dfr_dist_series}
\alias{is_dfr_dist_series}
\alias{ncomponents}
\alias{component}
\alias{component_hazard}
\alias{param_layout}
\alias{sample_components}
\alias{dfr_exponential}
\alias{dfr_weibull}
\alias{dfr_gompertz}
\alias{dfr_loglogistic}
\alias{cum_haz}
\alias{hazard}
\alias{surv}
\alias{sampler}
\alias{params}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{algebraic.dist}{\code{\link[algebraic.dist]{hazard}}, \code{\link[algebraic.dist]{params}}, \code{\link[algebraic.dist]{sampler}}, \code{\link[algebraic.dist]{surv}}}

  \item{dfr.dist}{\code{\link[dfr.dist]{cum_haz}}, \code{\link[dfr.dist]{dfr_exponential}}, \code{\link[dfr.dist]{dfr_gompertz}}, \code{\link[dfr.dist]{dfr_loglogistic}}, \code{\link[dfr.dist]{dfr_weibull}}}

  \item{dfr.dist.series}{\code{\link[dfr.dist.series]{component}}, \code{\link[dfr.dist.series]{component_hazard}}, \code{\link[dfr.dist.series]{dfr_dist_series}}, \code{\link[dfr.dist.series]{is_dfr_dist_series}}, \code{\link[dfr.dist.series]{ncomponents}}, \code{\link[dfr.dist.series]{param_layout}}, \code{\link[dfr.dist.series]{sample_components}}}

  \item{generics}{\code{\link[generics]{fit}}}

  \item{likelihood.model}{\code{\link[likelihood.model]{assumptions}}, \code{\link[likelihood.model]{hess_loglik}}, \code{\link[likelihood.model]{loglik}}, \code{\link[likelihood.model]{rdata}}, \code{\link[likelihood.model]{score}}}
}}

