% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{censored_integral}
\alias{censored_integral}
\title{Compute integral for left/interval censored contributions}
\usage{
censored_integral(h_fns, H_fns, layout, par, C_i, lower, upper, m)
}
\arguments{
\item{h_fns}{list of hazard closures}

\item{H_fns}{list of cumulative hazard closures}

\item{layout}{parameter layout}

\item{par}{parameter vector}

\item{C_i}{logical vector of candidate components}

\item{lower}{lower integration bound}

\item{upper}{upper integration bound}

\item{m}{number of components}
}
\value{
numeric integral value
}
\description{
Integrates \verb{[sum_\{j in C_i\} h_j(t)] * S_sys(t)} over \verb{(lower, upper)}.
}
\keyword{internal}
