% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{component_hazard.dfr_series_md}
\alias{component_hazard.dfr_series_md}
\title{Component hazard for a masked-cause DFR series system}
\usage{
\method{component_hazard}{dfr_series_md}(x, j, ...)
}
\arguments{
\item{x}{A \code{\link{dfr_series_md}} object.}

\item{j}{Component index.}

\item{...}{Additional arguments passed to the closure.}
}
\value{
A closure computing component j's hazard.
}
\description{
Component hazard for a masked-cause DFR series system
}
