% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hess.R
\name{hess_loglik.dfr_series_md}
\alias{hess_loglik.dfr_series_md}
\title{Hessian of log-likelihood for masked-cause DFR series systems}
\usage{
\method{hess_loglik}{dfr_series_md}(model, ...)
}
\arguments{
\item{model}{A \code{\link{dfr_series_md}} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A function with signature \code{function(df, par, ...)} returning
the Hessian matrix.
}
\description{
Returns a Hessian function computed via numerical differentiation of the
log-likelihood using \code{\link[numDeriv]{hessian}}.
}
