% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_series_md.R
\name{print.dfr_series_md}
\alias{print.dfr_series_md}
\title{Print method for dfr_series_md}
\usage{
\method{print}{dfr_series_md}(x, ...)
}
\arguments{
\item{x}{A \code{dfr_series_md} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for dfr_series_md
}
